Add createThread action and refactor startThread logic
