+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "itsmybot",
-  "version": "3.5.2",
+  "version": "3.5.4",
   "description": "",
   "main": "build/index.js",
   "scripts": {
